% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglabelr.R
\name{gglabelr}
\alias{gglabelr}
\title{Interactively make a label for a ggplot.}
\usage{
gglabelr(p)
}
\description{
When called, this function opens a window with options to interactively
annotate a ggplot. You can add a label and / or add a bounding box. You can
also view and copy the code to generate the annotations. When you press
the "done" button, the code to generate the annotations will be printed
to the console.
}
\examples{
library(gglabelr)
library(ggplot2)

# Make plot
p <- ggplot(mpg, aes(x = hwy, y = displ)) +
    geom_point()

# Interactively annotate the plot
gglabelr(p)
}
