% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBox.R
\name{makeBox}
\alias{makeBox}
\title{Interactively make a box for a ggplot.}
\usage{
makeBox(p)
}
\description{
When called, this function opens a window with options to interactively
add a box to the ggplot. When you press the "done" button, the code
to create the box will be printed to the console.
}
\examples{
library(gglabelr)
library(ggplot2)

# Make plot
p <- ggplot(mpg, aes(x = hwy, y = displ)) +
    geom_point()

# Interactively make a box
makeBox(p)
}
