% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabel.R
\name{makeLabel}
\alias{makeLabel}
\title{Interactively make a label for a ggplot.}
\usage{
makeLabel(p)
}
\description{
When called, this function opens a window with options to interactively
add a label to the ggplot. When you press the "done" button, the code
to create the label will be printed to the console.
}
\examples{
library(gglabelr)
library(ggplot2)

# Make plot
p <- ggplot(mpg, aes(x = hwy, y = displ)) +
    geom_point()

# Interactively make a label
makeLabel(p)
}
